update tests, have one demo service
