+++ projects/ui-components/lf-metadata/lf-field-template-container/lf-field-template-container-demo.service.ts
@@ -1,10 +1,14 @@
-// This service is an intentional duplication of the service in lf-documentation
-
-import { FieldType } from '@laserfiche/lf-ui-components/shared';
-import { FieldValues, TemplateFieldInfo, TemplateInfo } from '../field-components/utils/lf-field-types';
-import { isDynamicField } from '../field-components/utils/metadata-utils';
+import { FieldValues, LfFieldInfo, TemplateFieldInfo, TemplateInfo } from './../../../../ui-components/lf-metadata/field-components/utils/lf-field-types';
+import { LfFieldTemplateContainerService } from './../../../../ui-components/lf-metadata/lf-field-template-container/public-api';
+import { FieldType } from './../../../../ui-components/shared/lf-shared-public-api';
 import { DatetimeUtils } from './lf-field-template-container-datetime-utils';
-import { LfFieldTemplateContainerService } from './lf-field-template-container.service';
+
+
+export function isDynamicField(fieldInfo: LfFieldInfo): boolean {
+  const fieldInfoAsTemplateInfo = fieldInfo as TemplateFieldInfo;
+  return fieldInfoAsTemplateInfo.rule?.ancestors !== undefined;
+}
+
 
 export enum TemplateIds {
   String,
@@ -15,7 +19,8 @@ export enum TemplateIds {
   DynamicOneLocation,
   DynamicDateTimes,
   DynamicNumbers,
-  Groups
+  Groups,
+  Error
 }
 
 export enum DynamicFieldIds {
@@ -64,19 +69,19 @@ export enum City {
 }
 export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] } = {
   [TemplateIds.String]: [
-    { id: 1, name: 'multivalue string', fieldType: FieldType.String, isMultiValue: true },
-    { id: 2, name: 'plain string', fieldType: FieldType.String },
+    { id: 1, name: 'multivalue string', fieldType: FieldType.String, isMultiValue: true , displayName: 'multivalue string'},
+    { id: 2, name: 'plain string', fieldType: FieldType.String, displayName: 'plain string' },
   ],
   [TemplateIds.DateTime]: [
-    { id: 3, name: 'multivalue date', fieldType: FieldType.Date, isMultiValue: true },
-    { id: 4, name: 'multivalue datetime', fieldType: FieldType.DateTime, isMultiValue: true },
-    { id: 5, name: 'multivalue time', fieldType: FieldType.Time, isMultiValue: true },
+    { id: 3, name: 'multivalue date', fieldType: FieldType.Date, isMultiValue: true, displayName: 'multivalue date' },
+    { id: 4, name: 'multivalue datetime', fieldType: FieldType.DateTime, isMultiValue: true, displayName: 'multivalue datetime' },
+    { id: 5, name: 'multivalue time', fieldType: FieldType.Time, isMultiValue: true, displayName: 'multivalue time' },
   ],
   [TemplateIds.Number]: [
-    { id: 6, name: 'number', fieldType: FieldType.Number },
-    { id: 7, name: 'short integer 1', fieldType: FieldType.ShortInteger },
-    { id: 8, name: 'long interger', fieldType: FieldType.LongInteger },
-    { id: 9, name: 'short integer 2', fieldType: FieldType.ShortInteger },
+    { id: 6, name: 'number', fieldType: FieldType.Number, displayName: 'number' },
+    { id: 7, name: 'short integer 1', fieldType: FieldType.ShortInteger, displayName: 'short integer 1' },
+    { id: 8, name: 'long interger', fieldType: FieldType.LongInteger, displayName: 'long interger' },
+    { id: 9, name: 'short integer 2', fieldType: FieldType.ShortInteger, displayName: 'short integer 2' },
   ],
   [TemplateIds.DynamicLocation]: [
     {
@@ -85,7 +90,8 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       fieldType: FieldType.String,
       rule: { ancestors: [] },
       groupId: 1,
-      isMultiValue: true
+      isMultiValue: true,
+      displayName: 'dynamic state'
     },
     {
       id: DynamicFieldIds.County,
@@ -93,7 +99,8 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       fieldType: FieldType.String,
       rule: { ancestors: [DynamicFieldIds.State] },
       groupId: 1,
-      isMultiValue: true
+      isMultiValue: true,
+      displayName: 'dynamic county'
     },
     {
       id: DynamicFieldIds.City,
@@ -101,12 +108,14 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       fieldType: FieldType.String,
       rule: { ancestors: [DynamicFieldIds.County] },
       groupId: 1,
-      isMultiValue: true
+      isMultiValue: true,
+      displayName: 'dynamic city'
     },
     {
       id: 1000,
       name: 'Weather',
-      fieldType: FieldType.String
+      fieldType: FieldType.String,
+      displayName: 'weather'
     },
   ],
   [TemplateIds.DynamicOneLocation]: [
@@ -115,18 +124,21 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       name: 'State',
       fieldType: FieldType.String,
       rule: { ancestors: [] },
+      displayName: 'dynamic state'
     },
     {
       id: DynamicFieldIds.County,
       name: 'County',
       fieldType: FieldType.String,
       rule: { ancestors: [DynamicFieldIds.State] },
+      displayName: 'dynamic county'
     },
     {
       id: DynamicFieldIds.City,
       name: 'City',
       fieldType: FieldType.String,
       rule: { ancestors: [DynamicFieldIds.County] },
+      displayName: 'dynamic city'
     },
   ],
   [TemplateIds.DynamicDateTimes]: [
@@ -135,54 +147,63 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       name: 'Grandparent Date',
       fieldType: FieldType.Date,
       rule: { ancestors: [] },
+      displayName: 'dynamic grandparent'
     },
     {
       id: DynamicFieldIds.ParentDate,
       name: 'Parent Date',
       fieldType: FieldType.Date,
       rule: { ancestors: [DynamicFieldIds.GrandparentDate] },
+      displayName: 'dynamic parent date'
     },
     {
       id: DynamicFieldIds.ChildDate,
       name: 'Child Date',
       fieldType: FieldType.Date,
       rule: { ancestors: [DynamicFieldIds.ParentDate] },
+      displayName: 'dynamic child date'
     },
     {
       id: DynamicFieldIds.GrandparentTime,
       name: 'Grandparent Time',
       fieldType: FieldType.Time,
       rule: { ancestors: [] },
+      displayName: 'dynamic grandparenttime'
     },
     {
       id: DynamicFieldIds.ParentTime,
       name: 'Parent Time',
       fieldType: FieldType.Time,
       rule: { ancestors: [DynamicFieldIds.GrandparentTime] },
+      displayName: 'dynamic parent time'
     },
     {
       id: DynamicFieldIds.ChildTime,
       name: 'Child Time',
       fieldType: FieldType.Time,
       rule: { ancestors: [DynamicFieldIds.ParentTime] },
+      displayName: 'dynamic child time'
     },
     {
       id: DynamicFieldIds.GrandparentDateTime,
       name: 'Grandparent DateTime',
       fieldType: FieldType.DateTime,
       rule: { ancestors: [] },
+      displayName: 'grandparent datetime'
     },
     {
       id: DynamicFieldIds.ParentDateTime,
       name: 'Parent DateTime',
       fieldType: FieldType.DateTime,
       rule: { ancestors: [DynamicFieldIds.GrandparentDateTime] },
+      displayName: 'parent datetime'
     },
     {
       id: DynamicFieldIds.ChildDateTime,
       name: 'Child DateTime',
       fieldType: FieldType.DateTime,
       rule: { ancestors: [DynamicFieldIds.ParentDateTime] },
+      displayName: 'child datetime'
     },
   ],
   [TemplateIds.DynamicNumbers]: [
@@ -191,12 +212,14 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       name: 'Parent Number',
       fieldType: FieldType.Number,
       rule: { ancestors: [] },
+      displayName: 'dynamic parent num'
     },
     {
       id: DynamicFieldIds.ChildNumber,
       name: 'Child Number',
       fieldType: FieldType.Number,
       rule: { ancestors: [DynamicFieldIds.ParentNumber] },
+      displayName: 'child number'
     },
   ],
   [TemplateIds.Groups]: [
@@ -206,7 +229,8 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       fieldType: FieldType.String,
       groupId: 1,
       isMultiValue: true,
-      isRequired: true
+      isRequired: true,
+      displayName: 'group breakfast'
     },
     {
       id: 11,
@@ -214,39 +238,44 @@ export const fieldInfosPerTemplate: { [templateId: number]: TemplateFieldInfo[] 
       fieldType: FieldType.String,
       groupId: 1,
       isMultiValue: true,
-      defaultValue: 'hamburger'
+      defaultValue: 'hamburger',
+      displayName: 'group lunch'
     },
     {
       id: 12,
       name: 'Dinner',
       fieldType: FieldType.String,
       groupId: 1,
-      isMultiValue: true
+      isMultiValue: true,
+      displayName: 'group dinner'
     },
     {
       id: 13,
       name: 'Dessert',
       fieldType: FieldType.String,
-      isMultiValue: true
+      isMultiValue: true,
+      displayName: 'group dessert'
     },
     {
       id: 14,
       name: 'Snack',
-      fieldType: FieldType.String
+      fieldType: FieldType.String,
+      displayName: 'group snack'
     }
   ]
 };
 
 export class LfFieldTemplateContainerDemoService implements LfFieldTemplateContainerService {
   readonly testTemplateInfos: TemplateInfo[] = [
-    { id: TemplateIds.String, name: 'Strings' },
-    { id: TemplateIds.DateTime, name: 'Dates and Times' },
-    { id: TemplateIds.Number, name: 'Numbers' },
-    { id: TemplateIds.DynamicLocation, name: 'Location (Dynamic)' },
-    { id: TemplateIds.DynamicOneLocation, name: 'One Location (Dynamic)' },
-    { id: TemplateIds.DynamicDateTimes, name: 'Dates and Times (Dynamic)' },
-    { id: TemplateIds.DynamicNumbers, name: 'One Row of Numbers (Dynamic)' },
-    { id: TemplateIds.Groups, name: 'Template with Groups' }
+    { id: TemplateIds.String, name: 'Strings', displayName: 'strings template' },
+    { id: TemplateIds.DateTime, name: 'Dates and Times', displayName: 'dates and times template' },
+    { id: TemplateIds.Number, name: 'Numbers', displayName: 'numbers template' },
+    { id: TemplateIds.Error, name: 'Template with Error', displayName: 'Template With Error' },
+    { id: TemplateIds.DynamicLocation, name: 'Location (Dynamic)', displayName: 'Location (Dynamic)' },
+    { id: TemplateIds.DynamicOneLocation, name: 'One Location (Dynamic)', displayName: 'One Location (Dynamic)' },
+    { id: TemplateIds.DynamicDateTimes, name: 'Dates and Times (Dynamic)', displayName: 'Dates and Times (Dynamic)' },
+    { id: TemplateIds.DynamicNumbers, name: 'One Row of Numbers (Dynamic)', displayName: 'One Row of Numbers (Dynamic)' },
+    { id: TemplateIds.Groups, name: 'Template with Groups', displayName: 'Template with Groups' }
   ];
 
   readonly lookupTableByTemplateId = new Map<TemplateIds, LookupTable>([
@@ -343,18 +372,37 @@ export class LfFieldTemplateContainerDemoService implements LfFieldTemplateConta
   ]);
 
   async getAvailableTemplatesAsync(): Promise<TemplateInfo[]> {
-    return this.testTemplateInfos;
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+       return resolve(this.testTemplateInfos);
+    }, 200);
+  });
   }
 
   async getTemplateFieldsAsync(templateId: number): Promise<TemplateFieldInfo[]> {
+    if (templateId === TemplateIds.Error) {
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+         return reject();
+        }, 200);
+      });
+      }
     if (templateId === 0 || !(templateId in fieldInfosPerTemplate)) {
       return [];
     }
-    return fieldInfosPerTemplate[templateId];
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+       return resolve(fieldInfosPerTemplate[templateId]);
+    }, 200);
+  });
   }
 
   async getTemplateDefinitionAsync(id: string | number): Promise<TemplateInfo | undefined> {
-    return this.testTemplateInfos.find((info) => info.id === id);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+       return resolve(this.testTemplateInfos.find((info) => info.id === id));
+    }, 200);
+  });
   }
 
   async getDynamicFieldValueOptionsAsync(
